rename string_processing->utterance.py and make lowercasing non-optional
