+++ azimuth/utils/utterance.py
@@ -3,16 +3,14 @@
 # in the root directory of this source tree.
 
 
-def clean_utterance(utterance: str, lowercase: bool = True):
+def clean_utterance(utterance: str):
     """Converts characters as appropriate for downstream functions (e.g., search, spaCy)
 
     Args:
         utterance (str): Utterance to be cleaned
-        lowercase (bool): Lowercase all characters?
 
     Returns:
         utterance_clean (str): Cleaned utterance
     """
-    utterance_clean = utterance.replace("’", "'")  # Single quote -> apostrophe (unicode 8217 -> 39)
-    utterance_clean = utterance_clean.lower() if lowercase else utterance_clean
+    utterance_clean = utterance.replace("’", "'").lower()  # Quote->apostrophe (unicode 8217 -> 39)
     return utterance_clean
